requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
langchain
httpx
pytest
